add duration plan

+++ app/models/charity.rb
@@ -5,24 +5,24 @@ class Charity < ActiveRecord::Base
 
 
   def self.count_cart(user)
-    sum = 0.0
+    sum = 0
     unless user == nil
     @cart = Donation.where(:user_id => user.id)
     @cart.each do |donation|
-      sum += donation.amount.to_f
+      sum += donation.amount.to_i
     end
   end
-    @amount = sum * 100.00
+    @amount = sum * 100
     "$" + sum.to_s
   end
 
   def self.count_cart_int(user)
-    sum = 0.0
+    sum = 0
     @cart = Donation.where(:user_id => user.id)
     @cart.each do |donation|
-      sum += donation.amount.to_f
+      sum += donation.amount.to_i
     end
-    @amount = sum * 100.00
+    @amount = sum * 100
   end
